chore(physics): move `AddPhysics` for `App `in a dedicated module
