+++ src/plugins/physics/src/app.rs
@@ -1,0 +1,1 @@
+pub(crate) mod add_physics;
